Strenghtened schema handling to self-adapt
Cleaned-up the legacy code and removed some complicating unneccessary
section - like fitting the row data into a smaller array , which saves
nothing in term of space, processing or simplicity!

Did away with it using a dictionary where we store the location of
supported fields.

Could do further improvment but this is not working well with production
data adding the RX/TX stats.